requirements.txt
python-telegram-bot==5.3.0
eutils==0.2.2
biopython==1.68
numpy==1.11.0
pytesseract==0.1.6
Pillow==3.1.2
requests==2.9.1
